Switched from the pseudorandom number generator in cstdlib to an implementation of the Mersenne Twister (see http://www.math.sci.hiroshima-u.ac.jp/~m-mat/MT/emt.html) which should produce a more uniform distribution.

+++ include/thir.h
@@ -49,4 +49,4 @@
 LRESULT CALLBACK dice_proc(HWND w, UINT msg, WPARAM wp, LPARAM lp);
 void AlterEdit(HWND w, UINT id, int mod);
 void AlterEditU(HWND w, UINT id, int mod);
-UINT roll(UINT sides);
+unsigned long roll(UINT sides);
